Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/script-src-in-iframe-expected.txt

CONSOLE MESSAGE: The 'allow' directive has been replaced with 'default-src'. Please use that directive instead, as 'allow' has no effect.
CONSOLE MESSAGE: The 'allow' directive has been replaced with 'default-src'. Please use that directive instead, as 'allow' has no effect.
Loads an iframe (a) which loads an iframe (b) which in turns tries to load an external script. The iframe (a) has a content security policy disabling external scripts. As this policy does not apply to (b), the script should be executed.



--------
Frame: '<!--frame1-->'
--------


--------
Frame: '<!--frame2-->'
--------
PASS
